+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added support for dark mode
 - Added version number to the 'About' page
+- Fix a bug when running Find Your Place for the first time.
 
 ## 0.0.1